+++ src/compute_functions.jl
@@ -2,11 +2,12 @@
 """
     compute_trans_in(m, t, p, tm::TransmissionMode)
 
-Return the amount of resources going into transmission corridor l by a generic transmission mode.
+Return the amount of resources going into transmission corridor `l` by a generic
+`TransmissionMode`
 """
 function compute_trans_in(m, t, p, tm::TransmissionMode)
     exp = 0
-    if tm.Resource == p
+    if tm.resource == p
         exp += m[:trans_in][tm, t]
     end
     return exp
@@ -15,15 +16,15 @@ end
 """
     compute_trans_in(m, t, p, tm::PipeMode)
 
-Return the amount of resources going into transmission corridor l by a PipeMode transmission mode.
+Return the amount of resources going into transmission corridor `l` by a `PipeMode`.
 """
 function compute_trans_in(m, t, p, tm::PipeMode)
     exp = 0
-    if tm.Inlet == p
+    if inputs(tm)[1] == p
         exp += m[:trans_in][tm, t]
     end
-    if tm.Consuming == p
-        exp += m[:trans_in][tm, t] * tm.Consumption_rate[t]
+    if inputs(tm)[2] == p
+        exp += m[:trans_in][tm, t] * consumption_rate(tm, t)
     end
     return exp
 end
@@ -31,11 +32,12 @@ end
 """
     compute_trans_out(m, t, p, tm::TransmissionMode)
 
-Return the amount of resources going out of transmission corridor l by a generic transmission mode.
+Return the amount of resources going out of transmission corridor `l` by a generic
+`TransmissionMode`
 """
 function compute_trans_out(m, t, p, tm::TransmissionMode)
     exp = 0
-    if tm.Resource == p
+    if inputs(tm)[1] == p
         exp += m[:trans_out][tm, t]
     end
     return exp
@@ -44,11 +46,11 @@ end
 """
     compute_trans_out(m, t, p, tm::PipeMode)
 
-Return the amount of resources going out of transmission corridor l by a PipeMode transmission mode.
+Return the amount of resources going out of transmission corridor `l` by a `PipeMode`.
 """
 function compute_trans_out(m, t, p, tm::PipeMode)
     exp = 0
-    if tm.Outlet == p
+    if outputs(tm)[1] == p
         exp += m[:trans_out][tm, t]
     end
     return exp